// src/types/partials/VerificationResultAttestation.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TReference } from '../partials/Reference';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TDate } from '../simpleTypes/Date';
import { TSignature } from '../partials/Signature';

export type TVerificationResultAttestation = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    who?: TReference;
    onBehalfOf?: TReference;
    communicationMethod?: TCodeableConcept;
    date?: TDate;
    sourceIdentityCertificate?: String;
    proxyIdentityCertificate?: String;
    proxySignature?: TSignature;
    sourceSignature?: TSignature;
};